+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Removed `DateTime::operator Duration()`.
 - Moved `Azure::Core::BearerTokenAuthenticationPolicy`, defined in `azure/core/credentials.hpp` to `Azure::Core::Http` namespace in `azure/core/http/policy.hpp` header.
+- Removed option `AllowBeast` from `CurlTransportSSLOptions` in `CurlTransportOptions`.
 - Changed default option `NoRevoke` from `CurlTransportSSLOptions` for the `CurlTransportOptions` to `true`. This disables the revocation list checking by default.
 
 ## 1.0.0-beta.3 (2020-11-11)